backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
pydantic==2.9.2
python-multipart==0.0.12
aiofiles==24.1.0
httpx==0.27.2
websockets==13.1
aiohttp==3.10.10
pillow==10.4.0
